+++ tests/Rise.js
@@ -15,25 +15,47 @@ describe('Rise', function() {
         new Rise(document.createElement('div')).should.be.an.instanceof(Rise).and.be.ok;
     });
 
-    it('Should properly return parent node', function() {
+    it('Should properly call update method', function() {
+        var rise = new Rise('#rise-test');
+        rise.update().should.be.an.instanceof(Rise);
+    });
+
+    it('Should properly get/set parent node', function() {
         var rise = new Rise('#rise-test');
 
         rise.getParentNode().should.be.an.instanceof(Rise.RQuery);
         rise.getParentNode().is('div').should.be.ok;
+
+        rise.setParentNode(Rise.$.create('a')).should.be.an.instanceof(Rise);
+        rise.getParentNode().should.be.an.instanceof(Rise.RQuery);
+        rise.getParentNode().is('a').should.be.ok;
     });
 
-    it('Should properly return canvas node', function() {
+    it('Should properly get/set canvas node', function() {
         var rise = new Rise('#rise-test');
 
         rise.getCanvasNode().should.be.an.instanceof(Rise.RQuery);
         rise.getCanvasNode().is('div').should.be.ok;
+
+        rise.setCanvasNode(Rise.$.create('span')).should.be.an.instanceof(Rise);
+        rise.getCanvasNode().should.be.an.instanceof(Rise.RQuery);
+        rise.getCanvasNode().is('span').should.be.ok;
     });
 
     it('Should properly get/set config', function() {
-        var rise = new Rise('#rise-test');
+        var rise = new Rise('#rise-test', {
+            foo: 'test',
+            bar: {
+                key: 'value'
+            }
+        });
+
+        rise.getConfig('foo').should.be.equal('test');
+        rise.getConfig('bar.key').should.be.equal('value');
 
         rise.setConfig({
-            foo: 'bar',
+            foo: 'bar'
+        }, {
             bar: {
                 foo: 'bar'
             }
@@ -41,10 +63,12 @@ describe('Rise', function() {
 
         rise.getConfig('foo').should.be.equal('bar');
         rise.getConfig('bar.foo').should.be.equal('bar');
+        rise.getConfig('bar.key').should.be.equal('value');
         rise.getConfig().should.containEql({
             foo: 'bar',
             bar: {
-                foo: 'bar'
+                foo: 'bar',
+                key: 'value'
             }
         });
     });
@@ -75,6 +99,29 @@ describe('Rise', function() {
             width: 200,
             height: 200
         });
+    });
+
+    it('Should properly get/set HTML', function() {
+        var rise = new Rise('#rise-test');
+
+        rise.getHtml().should.be.a.string;
+        rise.setHtml('<span>Test</span>').should.be.an.instanceof(Rise);
+        rise.getHtml().should.be.equal('<span>Test</span>');
+
+        rise.getCanvasNode().should.be.an.instanceof(Rise.RQuery);
+        rise.getCanvasNode().is('span').should.be.ok;
+        rise.getParentNode().find('span').count().should.be.equal(1);
+    });
+
+    it('Should properly add new Element', function() {
+        var rise = new Rise('#rise-test'),
+            element = new Rise.Element();
+
+        rise.getCanvasNode().children().count().should.be.equal(0);
+        rise.addElement(element).should.be.an.instanceof(Rise);
+        rise.getCanvasNode().children().count().should.be.equal(1);
+        rise.getCanvasNode().empty().should.be.an.instanceof(Rise.RQuery);
+        rise.getCanvasNode().children().count().should.be.equal(0);
     });
 
     it('Should properly return current version', function() {